+++ CHANGELOG.md
@@ -9,6 +9,10 @@
 * frontend
   * Added sub menu for admin links
 
+### May 13, 2022
+* backend
+  * added new migrations and seeds for picklists database table
+
 ### May 6, 2022
 * DESCW-271 - improve linter standards
   * frontend and backend